Going in the right direction
Updated the fitness function to include not just the total distance traveled, but also the distance towards a target.

Updated the realtime_from_csv.py environment to be the new sandbox with mountain environment for the midterm.

Updated the simulation.py environment used to generate and train creatures to be the same new environment.

Moved output files such as .csv files and .urdf files to their respective output directories to keep the main folder structured.